hw-jv-oop-advanced/fix bugs

+++ src/main/java/core/basesyntax/IsoscelesTrapezoid.java
@@ -1,11 +1,11 @@
 package core.basesyntax;
 
 public class IsoscelesTrapezoid extends Figure {
-    private int firstSide;
-    private int secondSide;
-    private int height;
+    private double firstSide;
+    private double secondSide;
+    private double height;
 
-    public IsoscelesTrapezoid(String color, int firstSide, int secondSide, int height) {
+    public IsoscelesTrapezoid(String color, double firstSide, double secondSide, double height) {
         super(color);
         this.firstSide = firstSide;
         this.secondSide = secondSide;
