+++ src/main/java/core/basesyntax/Rectangle.java
@@ -1,10 +1,10 @@
 package core.basesyntax;
 
 public class Rectangle extends Figure {
-    private int firstSide;
-    private int secondSide;
+    private double firstSide;
+    private double secondSide;
 
-    public Rectangle(String color, int firstSide, int secondSide) {
+    public Rectangle(String color, double firstSide, double secondSide) {
         super(color);
         this.firstSide = firstSide;
         this.secondSide = secondSide;
